combined images to one large image
- IE supports images of width/height max ~16k, Firefox max 32k (see
http://www.osric.com/chris/verytallimages.html)
- both errors and figures directory contain the script
combineImages.py. They are slightly different versions
- combineImages.py requires the PIL module
- run combineImages.py on python 3
- errors-1/combineImages.py was a failed attempt to create 4 rows of
images, and then concatenate them top-to-bottom to create a more square
image. It failed because one of the images is slightly larger than the
others, and the program requires that all images be the same size,
otherwise it will warp the rest of the images